requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
passlib[bcrypt]
python-multipart
python-docx
pdfminer.six
reportlab
python-dotenv
